raop: Use sofware volume control, set device volume to -144.0 on connect
First time I got connected to my media renderer, I noticed the volume
was maxed-out on the media renderer. Turns out it was because the output
device left/right channels were set at 100% by default, which corresponds
to 0.00dB attenuation in Airtunes2 specifications [1], that beeing
equals to the media renderer maximum volume.

Note that I had no problem of bad volume computations [2]. Muting the
device was correctly sending an RTSP SET_PARAMETER message with volume
at -144.0, and changing the volume sliders in pavucontrol made the
SET_PARAMETER range from -30.0 to 0.00.

However I believe the current way of handling volume is potentially too
dangerous for speakers. My idea was to enable pulseaudio software volume
on the output device so that effetive HW volume won't be changed at all
on the media renderer. To do that I removed the set volume/mute callbacks
initialization and prevented any SET_PARAMETER volume RTSP message to be
sent out to the media renderer (including the one sent at connect time).

The RAOP2 specifications seams to indicate that the initial volume MUST
be sent when establishing a connection with a media renderer (see [3] and
[1]). On my Denon AVR-2113, not sending the initial SET_PARAMETER (volume)
at all makes it raising volume automatically to the amp's maximum: 74.0
on the LCD display, which corresponds to pulseaudio's 0.0 dB (100%). The
Denon may be using some kind of initial or default value to 0.0, which
is too bad because if you are not quick enough to turn off everything: it
may damage you ears, your speakers and the relationship with your
neighbors ;-)

After what I saw on the Denon amp, we can indeed assume that the inital
SET_PARAMETER volume is required. The Airtunes2 specifications should
probably make it optional, or, should implement a GET_PARAMETER volume
as a MUST HAVE. The GET_PARAMETER volume would enable pulseaudio to send
an initial SET_PARAMETER volume value so that the amp's keeps its current
volume. It would also allow pulseaudio to base its 100% volume on that
value. This should probably be an optional behavior, but it would be the
safest on amp's devices like my Denon. In the meantime, that patch sends
an initial SET_PARAMETER volume of -144.0.

I am not really happy with the connect process is done because I have
no way to make sure the initial SET_PARAMETER volume -144.0 was received
and processed correctly by the media renderer. It would probably require
some modification in the IO handling code. I am looking for a way to
properly handle a that scenario. If the RTSP/1.0 200 OK coresponding to
the initial (connect time) SET_PARAMETER volume -144.0 is not received,
then I would like to find a way to properly to kill the sink, without
killing the pulseaudio process (without an assertion).

[1] http://git.zx2c4.com/Airtunes2/about/#setting-volume
[2] https://bugs.freedesktop.org/show_bug.cgi?id=42804#c43
[3] http://git.zx2c4.com/Airtunes2/about/#connect