raop: Add support for module option protect_speakers
When enabled, this option disables hardware volume control so that
software volume control is used instead. In addition, it sets the initial
hardware volume muted on connect. The option is not enabled by default.

First time I got RAOP2/UDP working on my Denon AVR-2113, I noticed the
volume got maxed-out on the hardware. Turns out it was because the output
device left/right channels were set at 100% by default. That behavior
can be very bad for the speakers and the neightbors if you live in an
apartment!

I believe there are two kinds of setups:

- The output device is an audio amplifier with speakers directly
connected to it. That's the setup I have at home with my Denon.

- The output device is connected as a source for an external amplifier.
This kind of setup is fine because the volume level on the external
amplifier does not change, removing the risk of burning your speakers.

Even though the Airtune2 specification available at
http://git.zx2c4.com/Airtunes2/about/ states that an RTSP SET_PARAMETER
request should be sent after the RTSP RECORD request in order to set the
initial volume, I first tried NOT to send it at connect time. The
amplifier raised the volume to its maximum... I don't know if that
behavior is expected or if it may be a bug on the Denon because of a
variable initialized to 0.00. I don't know either how other devices
behave if no initial RTSP SET_PARAMETER is sent regarding the volume.

The protect_speakers option intends to prevent max-volume mistakes when
using RAOP on standalone audio amplifiers devices. When enabled, the
initial volume is muted on connect and the volume/mute callbacks are set
to NULL, thus enabling software volume control for the RAOP output
device. It should be noted that the result of the initial RTSP
SET_PARAMETER (volume) result is not checked at the moment, so using an
audio amplifier occasionally as RAOP device might still be dangerous for
the speakers even with that option enabled. If the RTSP/1.0 200 OK
coresponding to the initial RTSP SET_PARAMETER volume -144.0 (mute) is
not received, then it would be great to stop the sink properly and
cancel the operation. The audio stream should start only after the
initial RTSP SET_PARAMETER (volume) response has been received and was
successful.